Move factories to handler classes
