+++ android/src/main/java/com/swmansion/gesturehandler/core/ManualGestureHandler.kt
@@ -1,11 +1,24 @@
 package com.swmansion.gesturehandler.core
 
+import android.content.Context
 import android.view.MotionEvent
+import com.swmansion.gesturehandler.react.eventbuilders.ManualGestureHandlerEventDataBuilder
 
 class ManualGestureHandler : GestureHandler<ManualGestureHandler>() {
   override fun onHandle(event: MotionEvent, sourceEvent: MotionEvent) {
     if (state == STATE_UNDETERMINED) {
       begin()
     }
+  }
+
+  class Factory : GestureHandler.Factory<ManualGestureHandler>() {
+    override val type = ManualGestureHandler::class.java
+    override val name = "ManualGestureHandler"
+
+    override fun create(context: Context?): ManualGestureHandler {
+      return ManualGestureHandler()
+    }
+
+    override fun createEventBuilder(handler: ManualGestureHandler) = ManualGestureHandlerEventDataBuilder(handler)
   }
 }
