+++ android/src/main/java/com/swmansion/gesturehandler/core/PinchGestureHandler.kt
@@ -1,8 +1,10 @@
 package com.swmansion.gesturehandler.core
 
+import android.content.Context
 import android.graphics.PointF
 import android.view.MotionEvent
 import android.view.ViewConfiguration
+import com.swmansion.gesturehandler.react.eventbuilders.PinchGestureHandlerEventDataBuilder
 import kotlin.math.abs
 
 class PinchGestureHandler : GestureHandler<PinchGestureHandler>() {
@@ -99,5 +101,16 @@ class PinchGestureHandler : GestureHandler<PinchGestureHandler>() {
   override fun resetProgress() {
     velocity = 0.0
     scale = 1.0
+  }
+
+  class Factory : GestureHandler.Factory<PinchGestureHandler>() {
+    override val type = PinchGestureHandler::class.java
+    override val name = "PinchGestureHandler"
+
+    override fun create(context: Context?): PinchGestureHandler {
+      return PinchGestureHandler()
+    }
+
+    override fun createEventBuilder(handler: PinchGestureHandler) = PinchGestureHandlerEventDataBuilder(handler)
   }
 }
