+++ android/src/main/java/com/swmansion/gesturehandler/core/TapGestureHandler.kt
@@ -1,10 +1,14 @@
 package com.swmansion.gesturehandler.core
 
+import android.content.Context
 import android.os.Handler
 import android.os.Looper
 import android.view.MotionEvent
+import com.facebook.react.bridge.ReadableMap
+import com.facebook.react.uimanager.PixelUtil
 import com.swmansion.gesturehandler.core.GestureUtils.getLastPointerX
 import com.swmansion.gesturehandler.core.GestureUtils.getLastPointerY
+import com.swmansion.gesturehandler.react.eventbuilders.TapGestureHandlerEventDataBuilder
 import kotlin.math.abs
 
 class TapGestureHandler : GestureHandler<TapGestureHandler>() {
@@ -160,6 +164,52 @@ class TapGestureHandler : GestureHandler<TapGestureHandler>() {
     tapsSoFar = 0
     currentMaxNumberOfPointers = 0
     handler?.removeCallbacksAndMessages(null)
+  }
+
+  class Factory : GestureHandler.Factory<TapGestureHandler>() {
+    override val type = TapGestureHandler::class.java
+    override val name = "TapGestureHandler"
+
+    override fun create(context: Context?): TapGestureHandler {
+      return TapGestureHandler()
+    }
+
+    override fun configure(handler: TapGestureHandler, config: ReadableMap) {
+      super.configure(handler, config)
+      if (config.hasKey(KEY_TAP_NUMBER_OF_TAPS)) {
+        handler.setNumberOfTaps(config.getInt(KEY_TAP_NUMBER_OF_TAPS))
+      }
+      if (config.hasKey(KEY_TAP_MAX_DURATION_MS)) {
+        handler.setMaxDurationMs(config.getInt(KEY_TAP_MAX_DURATION_MS).toLong())
+      }
+      if (config.hasKey(KEY_TAP_MAX_DELAY_MS)) {
+        handler.setMaxDelayMs(config.getInt(KEY_TAP_MAX_DELAY_MS).toLong())
+      }
+      if (config.hasKey(KEY_TAP_MAX_DELTA_X)) {
+        handler.setMaxDx(PixelUtil.toPixelFromDIP(config.getDouble(KEY_TAP_MAX_DELTA_X)))
+      }
+      if (config.hasKey(KEY_TAP_MAX_DELTA_Y)) {
+        handler.setMaxDy(PixelUtil.toPixelFromDIP(config.getDouble(KEY_TAP_MAX_DELTA_Y)))
+      }
+      if (config.hasKey(KEY_TAP_MAX_DIST)) {
+        handler.setMaxDist(PixelUtil.toPixelFromDIP(config.getDouble(KEY_TAP_MAX_DIST)))
+      }
+      if (config.hasKey(KEY_TAP_MIN_POINTERS)) {
+        handler.setMinNumberOfPointers(config.getInt(KEY_TAP_MIN_POINTERS))
+      }
+    }
+
+    override fun createEventBuilder(handler: TapGestureHandler) = TapGestureHandlerEventDataBuilder(handler)
+
+    companion object {
+      private const val KEY_TAP_NUMBER_OF_TAPS = "numberOfTaps"
+      private const val KEY_TAP_MAX_DURATION_MS = "maxDurationMs"
+      private const val KEY_TAP_MAX_DELAY_MS = "maxDelayMs"
+      private const val KEY_TAP_MAX_DELTA_X = "maxDeltaX"
+      private const val KEY_TAP_MAX_DELTA_Y = "maxDeltaY"
+      private const val KEY_TAP_MAX_DIST = "maxDist"
+      private const val KEY_TAP_MIN_POINTERS = "minPointers"
+    }
   }
 
   companion object {
